requirements.txt
Flask==3.0.3
tensorflow==2.12.0
numpy==1.23.5
opencv-python-headless==4.10.0.84
Pillow==10.4.0
h5py==3.11.0
